Converted everything to double

+++ final/Foundation4/Activity.cs
@@ -1,10 +1,10 @@
 public abstract class Activity
 {
     protected string _date;
-    protected int _length;
+    protected double _length;
     protected string _activityType;
 
-    public Activity(string date, int length)
+    public Activity(string date, double length)
     {
         _date = date;
         _length = length;
